# macrolens_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MacroLens API",
    "description": "宏观数据、发布日历、FOMC、文档检索和 AI 研究平台 API。",
    "version": "1.0.1"
  },
  "paths": {
...
    "/api/v1/release-events": {
      "get": {
        "tags": [
          "Releases"
        ],
        "summary": "List Events",
        "operationId": "list_events_api_v1_release_events_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "US",
              "title": "Country"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "importance_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Importance Min"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Events Api V1 Release Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/release-events/{event_id}": {
      "get": {
        "tags": [
          "Releases"
        ],
        "summary": "Event Detail",
        "operationId": "event_detail_api_v1_release_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/release-events/{event_id}/forecasts": {
      "get": {
        "tags": [
          "Releases"
        ],
        "summary": "Event Forecasts",
        "operationId": "event_forecasts_api_v1_release_events__event_id__forecasts_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ForecastItem"
                  },
                  "title": "Response Event Forecasts Api V1 Release Events  Event Id  Forecasts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/release-events/{event_id}/market-reactions": {
      "get": {
        "tags": [
          "Releases"
        ],
        "summary": "Event Market Reactions",
        "operationId": "event_market_reactions_api_v1_release_events__event_id__market_reactions_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketReactionItem"
                  },
                  "title": "Response Event Market Reactions Api V1 Release Events  Event Id  Market Reactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForecastItem": {
        "properties": {
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "consensus_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consensus Value"
          },
          "median_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Value"
          },
          "high_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "High Value"
          },
          "low_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Value"
          },
          "respondent_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Respondent Count"
          },
          "provider_code": {
            "type": "string",
            "title": "Provider Code"
          }
        },
        "type": "object",
        "required": [
          "observed_at",
          "consensus_value",
          "median_value",
          "high_value",
          "low_value",
          "respondent_count",
          "provider_code"
        ],
        "title": "ForecastItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketReactionItem": {
        "properties": {
          "instrument_code": {
            "type": "string",
            "title": "Instrument Code"
          },
          "window_code": {
            "type": "string",
            "title": "Window Code"
          },
          "absolute_change": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Change"
          },
          "percent_change": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percent Change"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "type": "object",
        "required": [
          "instrument_code",
          "window_code",
          "absolute_change",
          "percent_change",
          "observed_at"
        ],
        "title": "MarketReactionItem"
      },
...
      "ReleaseEventDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title_zh": {
            "type": "string",
            "title": "Title Zh"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title En"
          },
          "country_code": {
            "type": "string",
            "title": "Country Code"
          },
          "reference_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Period"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "actual_released_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Released At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "importance_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance Score"
          },
          "release_type": {
            "type": "string",
            "title": "Release Type"
          },
          "provider_code": {
            "type": "string",
            "title": "Provider Code"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "official_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Official Url"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/ReleaseMetric"
            },
            "type": "array",
            "title": "Metrics"
          },
          "consensus_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consensus Value"
          },
          "source_timezone": {
            "type": "string",
            "title": "Source Timezone"
          },
          "forecasts": {
            "items": {
              "$ref": "#/components/schemas/ForecastItem"
            },
            "type": "array",
            "title": "Forecasts"
          },
          "market_reactions": {
            "items": {
              "$ref": "#/components/schemas/MarketReactionItem"
            },
            "type": "array",
            "title": "Market Reactions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title_zh",
          "title_en",
          "country_code",
          "reference_period",
          "scheduled_at",
          "actual_released_at",
          "status",
          "importance_score",
          "release_type",
          "provider_code",
          "provider_name",
          "official_url",
          "source_timezone"
        ],
        "title": "ReleaseEventDetail"
      },
      "ReleaseMetric": {
        "properties": {
          "series_id": {
            "type": "string",
            "format": "uuid",
            "title": "Series Id"
          },
          "name_zh": {
            "type": "string",
            "title": "Name Zh"
          },
          "transform": {
            "type": "string",
            "title": "Transform"
          },
          "actual_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Value"
          },
          "previous_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Value"
          },
          "revised_previous_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revised Previous Value"
          },
          "unit_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Label"
          }
        },
        "type": "object",
        "required": [
          "series_id",
          "name_zh",
          "transform",
          "actual_value",
          "previous_value",
          "revised_previous_value",
          "unit_label"
        ],
        "title": "ReleaseMetric"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}